app/requirements.txt
Django==5.0.6
psycopg2>=2.8
djangorestframework==3.15.2